add status change method to homeController and use it in dashboard

+++ database/migrations/2019_10_11_210358_create_user_menus_table.php
@@ -15,7 +15,7 @@ class CreateUserMenusTable extends Migration
     {
         Schema::create('user_menus', function (Blueprint $table) {
             $table->increments('id')->unsigned();
-            $table->tinyInteger('RoleId')->unsigned();
+            $table->Integer('RoleId')->unsigned();
             $table->string('MenuItem');
             $table->string('Url')->nullable();
             $table->timestamps();
